refactor(ocr): extract shared final review helpers
