+++ .github/scripts/final-ai-review.js
@@ -1,4 +1,14 @@
-const crypto = require('node:crypto')
+const {
+  getPermission,
+  listCheckRunsForRef,
+  repositoryName,
+  safeFence,
+  sha256,
+  validDigest,
+  validSha,
+  workflowRunIdFromUrl,
+  workflowRunUrl,
+} = require('./final-ai-review-shared')
 const { execFileSync } = require('node:child_process')
 const fs = require('node:fs')
 const os = require('node:os')
@@ -61,18 +71,6 @@ const FINDING_CATEGORIES = new Set([
 ])
 const FINDING_SEVERITIES = new Set(['critical', 'high', 'medium', 'low'])
 const DISPOSITION_STATES = new Set(['fixed', 'follow-up', 'rejected'])
-
-function sha256(value) {
-  return crypto.createHash('sha256').update(String(value)).digest('hex')
-}
-
-function validSha(value) {
-  return /^[0-9a-f]{40}$/.test(value ?? '')
-}
-
-function validDigest(value) {
-  return /^[0-9a-f]{64}$/.test(value ?? '')
-}
 
 function buildFinalReviewEvidenceDigest(value) {
   return sha256(JSON.stringify(value))
@@ -768,19 +766,6 @@ function validatePromotionContract(input) {
   }
 }
 
-function workflowRunUrl(context, runId = context.runId) {
-  return `${context.serverUrl}/${context.repo.owner}/${context.repo.repo}/actions/runs/${runId}`
-}
-
-function workflowRunIdFromUrl(context, targetUrl) {
-  const prefix = `${context.serverUrl}/${context.repo.owner}/${context.repo.repo}/actions/runs/`
-  if (typeof targetUrl !== 'string' || !targetUrl.startsWith(prefix)) {
-    return null
-  }
-  const value = targetUrl.slice(prefix.length)
-  return /^[1-9][0-9]*$/.test(value) ? Number(value) : null
-}
-
 async function setCommitStatus({ github, context, sha, state, description }) {
   await github.rest.repos.createCommitStatus({
     owner: context.repo.owner,
@@ -793,20 +778,6 @@ async function setCommitStatus({ github, context, sha, state, description }) {
   })
 }
 
-async function getPermission(github, context, username) {
-  try {
-    const response = await github.rest.repos.getCollaboratorPermissionLevel({
-      owner: context.repo.owner,
-      repo: context.repo.repo,
-      username,
-    })
-    return response.data.user?.permission ?? response.data.permission ?? ''
-  } catch (error) {
-    if (error.status === 404) return ''
-    throw error
-  }
-}
-
 async function getPull(github, context, pullNumber) {
   const response = await github.rest.pulls.get({
     owner: context.repo.owner,
@@ -814,10 +785,6 @@ async function getPull(github, context, pullNumber) {
     pull_number: pullNumber,
   })
   return response.data
-}
-
-function repositoryName(context) {
-  return `${context.repo.owner}/${context.repo.repo}`
 }
 
 function isEligibleDefaultPull({ pull, context, baseSha, headSha }) {
@@ -985,23 +952,6 @@ async function hasSuccessfulFinalReview(
     (workflowRunId == null ||
       status.target_url === workflowRunUrl(context, workflowRunId))
   )
-}
-
-async function listCheckRunsForRef({ github, context, ref }) {
-  if (
-    typeof github.paginate !== 'function' ||
-    typeof github.rest.checks?.listForRef !== 'function'
-  ) {
-    return []
-  }
-  const checkRuns = await github.paginate(github.rest.checks.listForRef, {
-    owner: context.repo.owner,
-    repo: context.repo.repo,
-    ref,
-    filter: 'all',
-    per_page: 100,
-  })
-  return Array.isArray(checkRuns) ? checkRuns : []
 }
 
 async function getVerifiedCleanFinalReviewEvidence({
@@ -2318,12 +2268,6 @@ async function startFinalReview({
   })
   core?.setOutput('background', plan.background)
   return true
-}
-
-function safeFence(value) {
-  const runs = String(value).match(/`+/g) ?? []
-  const longest = runs.reduce((max, run) => Math.max(max, run.length), 0)
-  return '`'.repeat(Math.max(3, longest + 1))
 }
 
 function fencedBlock(value) {
